+++ .github/FUNDING.yml
@@ -1,6 +1,6 @@
 # These are supported funding model platforms
 
-github: [mid0aria, vihaan reddy] # Replace with up to 4 GitHub Sponsors-enabled usernames e.g., [user1, user2]
+github: [mid0aria, vihaanreddy] # Replace with up to 4 GitHub Sponsors-enabled usernames e.g., [user1, user2]
 patreon: # Replace with a single Patreon username
 open_collective: # Replace with a single Open Collective username
 ko_fi: # Replace with a single Ko-fi username
