+++ changes.json
@@ -1,11 +1,12 @@
 {
     "c": [
         "bot.js",
-        "version.json",
+        ".gitignore",
+        "socket.bat",
         "package.json",
-        "phrases/phrases.json",
-        "utilfiles/captcha.png",
-        "socket.js",
-        "socket.bat"
+        "images/captcha.png",
+        "utilfiles/socket.js",
+        ".github/FUNDING.yml",
+        "blank.js"
     ]
 }
